+++ src/NSwag.Commands/HostFactoryResolver.cs
